### 2023-11-24 #6087 @editorialbot
[PRE REVIEW]: PyLogGrid: A Python package for fluid dynamics on logarithmic lattices
**Submitting author:** @hippalectryon-0 (<a href="http://orcid.org/0000-0002-5857-1290">Amaury Barral</a>)
**Repository:** https://github.com/hippalectryon-0/pyloggrid
**Branch with paper.md** (empty if default branch): joss
**Version:** 2.3.0
