### 2024-01-16 #6087 @philipcardiff
FYI, I invited Ciro S. Campolina to review this submission via email (and am awaiting a reply), since you refer to their LogLatt library.
