### 2024-01-20 #6087 @hippalectryon-0
@philipcardiff
Other than C. Campolina, I don't really have relevant reviewers from my field to suggest, as most aren't really in the numerical side of things.
Looking quickly are other JOSS fluid dynamics paper, no reviewers really ring a bell, so I'm not sure that I can make a very relevant suggestion. Just in case, here are a few reviewers that reviewed somewhat similar (?) fluid dynamics papers (I removed the "@" to not ping them all): ashwinvis, NoraLoose, ranocha, eviatarbach, DamienIrving
