### 2024-02-12 #6087 @philipcardiff
Hi @hippalectryon-0,

Would you mind addressing the following points on the documentation (https://pyloggrid.readthedocs.io/en/latest/benchmark.html#vs-grid-size):
- Fix missing images at https://pyloggrid.readthedocs.io/en/latest/benchmark.html#vs-grid-size
- Fix missing images at https://pyloggrid.readthedocs.io/en/latest/tutorial/tutorial_analyzing_output.html
- Fix typo at https://pyloggrid.readthedocs.io/en/latest/documentation/doc_solvers.html: *veru stiff* => *very stiff*
- Fix missing images at https://pyloggrid.readthedocs.io/en/latest/documentation/doc_performance.html
- Fix missing images at https://pyloggrid.readthedocs.io/en/latest/documentation/doc_plotting.html

Also, when I go to the main repository (https://github.com/hippalectryon-0/pyloggrid), the README file says *this is not the main repository* and *it may not always be up to date*; this is strange since the review is presumably meant to be on the main repository. At a minimum, please provide a link to the documentation in the README.

I am still searching for suitable reviewers.
