### 2024-02-12 #6087 @philipcardiff
For the tutorial at https://pyloggrid.readthedocs.io/en/latest/tutorial/tutorial_first_simu.html#running-the-simulation, can you add more details on the problem definition? e.g.
- Geometry
- Material properties
- Loading/boundary conditions
- Initial conditions

This would also help me find appropriate reviewers as I would better understand what type of problems PyLogGrid can solve.

Thanks.
