+++ README.md
@@ -6,6 +6,8 @@ The main package is currently hosted on a private Gitlab for pratical reasons.
 
 This repository is used to open-source the code and documentation, and allow issues and pull requests from the community.
 
+You can read the documentation [here](https://pyloggrid.readthedocs.io/).
+
 **The code of interest if located in log-grid/. The files in the root directory are only used to link the Gitlab and Github, and build documentation.**
 
 *Since this is not the main repository, and requires syncing by hand, it may not always be up to date with the latest version. If you think this repository is missing a version, don't hesitate to open an issue.*
