### 2024-02-13 #6087 @hippalectryon-0
@philipcardiff

> Would you mind addressing the following points on the documentation (https://pyloggrid.readthedocs.io/en/latest/benchmark.html#vs-grid-size): (...)

I've pushed a new version, with the fixed typo and missing images. To view it, **select the joss branch in the docs**, or go directly to https://pyloggrid.readthedocs.io/en/joss/index.html.

> Also, when I go to the main repository (https://github.com/hippalectryon-0/pyloggrid), the README file says this is not the main repository and it may not always be up to date; this is strange since the review is presumably meant to be on the main repository.

As explained elsewhere, the repository where the development takes place is a privately-hosted GitLab for administrative reasons, and the Github is simply a mirror to make it accessible to everyone.

> At a minimum, please provide a link to the documentation in the README.

Added

*Ps: thanks for all your work :)*
