### 2024-02-14 #6087 @philipcardiff
By email, C. Campolina let me know that he is available to review this submission; however, this would appear to be against the [JOSS Conflict of Interest Policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) as you have recently co-authored a [preprint](https://arxiv.org/pdf/2312.01702.pdf) together.

I will continue to search for another reviewer. Thanks for your patience.
