### 2024-02-29 #6087 @hippalectryon-0
@slaizet Can you try `git clone https://github.com/hippalectryon-0/pyloggrid.git` instead ?

(note that you'll then need to run `cd pyloggrid && git checkout joss` to switch to the joss branch)
