+++ log-grid/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.3]
+
+### Fixed
+
+- Fix github repo link in docs
+- Improve tutorial docs
+
 ## [2.3.2]
 
 ### Fixed
