### 2024-03-01 #6087 @slaizet
Hi @hippalectryon-0,

I am trying to run my "first simulation" following the instructions here: https://pyloggrid.readthedocs.io/en/latest/tutorial/tutorial_first_simu.html

First of all, you should specify all the Python libraries required to run `NS3D.py`. Secondly, I get an error message when trying to run the code with python:

```
File "NS3D.py", line 22
    def get_forcing(grid: Grid) -> tuple[np.ndarray, np.ndarray, np.ndarray]:
                        ^
SyntaxError: invalid syntax
```
I get a different error when trying to run the code with python3:

```
Traceback (most recent call last):
  File "NS3D.py", line 14, in <module>
    from pyloggrid.Libs.datasci import randcomplex_like, randcomplex_seeded_by_array
  File "/home/slaizet/reviewJOSS/pyloggrid/log-grid/Simulations/../pyloggrid/Libs/datasci.py", line 33, in <module>
    def mean(data: np.ndarray, ts: Optional[np.ndarray] = None, log: bool = False, start: float = 0, end: float = 1) -> tuple[np.ndarray, np.ndarray | None]:
TypeError: unsupported operand type(s) for |: 'type' and 'NoneType'
```
I can confirm that I am in the joss branch:

```
On branch joss
Your branch is up-to-date with 'origin/joss'.
```
Am I missing something?

Cheers,

Sylvain
