### 2024-03-01 #6087 @philipcardiff
Hi @marlonsmathias, Would you be interested in reviewing the submission “PyLogGrid: A Python package for fluid dynamics on logarithmic lattice”, which we received at JOSS? The link for the submission is https://github.com/openjournals/joss-reviews/issues/6087 , and the repository is at https://github.com/hippalectryon-0/pyloggrid.

If you are unfamiliar with JOSS, the review process is unique: it takes place in a GitHub issue, is open, and encourages author-reviewer-editor conversations. You can have a look at our review criteria (https://joss.readthedocs.io/en/latest/review_criteria.html) and review checklist (https://joss.readthedocs.io/en/latest/review_checklist.html) to get an idea of what is expected.
