### 2024-03-01 #6087 @marlonsmathias
Hello @philipcardiff , yes, I am interested in reviewing the submission. I have not reviewed for JOSS yet, so I will take a look at the documentation.
