### 2024-03-08 #6087 @hippalectryon-0
@slaizet
Apologies for the late answer, I missed the notification.

> First of all, you should specify all the Python libraries required to run NS3D.py. Secondly, I get an error message when trying to run the code with python:

Could you confirm that you have successfully ran `setup.sh` as indicated [here](https://pyloggrid.readthedocs.io/en/joss/tutorial/tutorial_installation.html) ? Your error indicates that you're using a python version which is too old. The setup takes care of installing the right python version, the virtual environment, and all the required packages.
