### 2024-03-16 #6439 @marlonsmathias
@philipcardiff I have found some issues in the installation process, but I'm unsure how to proceed. Do I report them in this issue or do I open a new issue?
