### 2024-04-08 #6439 @slaizet
Few comments about this submission:

Contribution and authorship:
It seems than only hippalectryon-0 & logseq7546515162 have contributed to Github. It would be good to clarify the role of the other authors in the paper.

Data sharing & Reproducibility:
It would be good to provide data for comparison for various examples. It will help with reproducibility.

Performance:
Performance data are missing (more than just Windows versus Linux). This can be done here https://pyloggrid.readthedocs.io/en/latest/documentation/doc_performance.html or in the paper.

Example usage:
More examples would be great.

Automated tests:
Missing.

State of the field:
Missing in the paper.
