### 2024-06-13 #6439 @hippalectryon-0
@marlonsmathias Thanks for your review !

- I see you've left the "performance" checkbox unchecked. Could you indicate more precisely which part is lacking ?
- Likewise, what more in "Functionality documentation" would you have liked ?
