### 2024-06-13 #6439 @hippalectryon-0
@slaizet Thanks for your review !

Regarding the points you have deemed lacking:

- "Contribution and authorship": could you indicate what is missing ?
- "Performance": same question
- "Example usage": could you indicate what is missing in the example that is provided in the documentation ?
- "Automated tests": I believe there are many automated tests in https://github.com/hippalectryon-0/pyloggrid/tree/joss/log-grid/tests
- "State of the field": Given that there are no other packages to make simulations on shell models other than the Matlab package by Campolina&Al that is mentioned in the doc, could you expand on what is missing ?

Best,
A.B.
