### 2024-07-03 #6439 @slaizet
Thank you for your reply.

I do not know if your answer regarding the contributions of the authors and having a single example of use of the software is satisfactory for the journal (@philipcardiff should be able to answer).

In terms of performance, it would be great to see some scalability plots, memory usage, etc on various hardware.

Finally, I still think that a small literature review on shell model simulations would be helpful for the readers and for potential users of your software.
