### 2024-07-03 #6439 @philipcardiff
> Thank you for your reply.
>
> I do not know if your answer regarding the contributions of the authors and having a single example of use of the software is satisfactory for the journal (@philipcardiff should be able to answer).
>
> In terms of performance, it would be great to see some scalability plots, memory usage, etc on various hardware.
>
> Finally, I still think that a small literature review on shell model simulations would be helpful for the readers and for potential users of your software.

Regarding authorship, the [JOSS documentation states](https://joss.readthedocs.io/en/latest/submitting.html#authorship) (note that authoring a commit is not a requirement):
>Purely financial (such as being named on an award) and organizational (such as general supervision of a research group) contributions are not considered sufficient for co-authorship of JOSS submissions, but active project direction and other forms of non-code contributions are. The authors themselves assume responsibility for deciding who should be credited with co-authorship, and co-authors must always agree to be listed. In addition, co-authors agree to be accountable for all aspects of the work, and to notify JOSS if any retraction or correction of mistakes are needed after publication.

@hippalectryon-0: based on your comments, I believe all five authors satisfy these requirements. Is this correct?

Regarding the number of example cases, the [JOSS documentation states](https://joss.readthedocs.io/en/latest/review_criteria.html#example-usage):
> The authors should include examples of how to use the software (ideally to solve real-world analysis problems).

Although possibly ambiguous, note the use of the word "example**s**" (plural).

Also, the [JOSS documentation states](https://joss.readthedocs.io/en/latest/review_criteria.html#substantial-scholarly-effort)
> Reviewers should rely on their expert understanding of their domain to judge whether the software is of broad interest (likely to be cited by other researchers) or more narrowly focused around the needs of an individual researcher or lab.

Based on these two points, it seems reasonable that more than one test case should be included in the documentation.
