### 2024-07-03 #6439 @marlonsmathias
> @marlonsmathias Thanks for your review !
>
> * I see you've left the "performance" checkbox unchecked. Could you indicate more precisely which part is lacking ?

Regarding the performance, it is now solved, I had left it unchecked because I could not find the data in the version of the documentation I was looking at (v:latest).

> * Likewise, what more in "Functionality documentation" would you have liked ?

For the functionality documentation, I am missing some information on the case setup and run in the tutorial. I would like to see more details on how to set up, for instance, the boundary conditions. One simple case for this is the lid-driven cavity, but others could also be used.
