### 2024-11-06 #6439 @hippalectryon-0
Apologies for the late answer, it is always difficult to find proper time to wrap up this paper. This was a software I developed for my PhD, but as I am now working somewhere else, I hardly have much time to work on this.
Nevertheless, seeing as this has dragged on for quite some time, I would like to wrap up and publish this piece of software before the end of the year if possible.

Looking back at the comments provided above, the only missing point is an additional example / test case in the documentation.
I will add a Homogeneous Rayleigh-Bénard case shortly.

To answer some questions asked above:
@philipcardiff
> @hippalectryon-0: based on your comments, I believe all five authors satisfy these requirements. Is this correct

This is correct.

@slaizet
> In terms of performance, it would be great to see some scalability plots, memory usage, etc on various hardware.

Although I understand that more stats may be useful, I believe that the current stats at https://pyloggrid.readthedocs.io/en/joss/benchmark.html are already sufficient, and moreover can easily be ran by users on their own hardware.
