### 2024-11-06 #6439 @hippalectryon-0
@marlonsmathias
> For the functionality documentation, I am missing some information on the case setup and run in the tutorial. I would like to see more details on how to set up, for instance, the boundary conditions.

I understand that this may seem nonintuitive for users of standard fluid simulation software, but in our case we do not prescribe any boundary conditions, as we are always in Fourier space, which implies periodic boundary conditions.
This is explained (shortly) in https://pyloggrid.readthedocs.io/en/joss/tutorial/tutorial_first_simu.html.
