### 2024-11-14 3a66714
improve docs, add HRB example
