+++ log-grid/docs/tutorial/tuto_HRB/tutorial_HRB.rst
@@ -1,0 +1,10 @@
+#######################################
+Tutorial 2: Homogeneous Rayleigh-Bénard
+#######################################
+
+.. warning:: This assumes you have completed :doc:`../tuto_basic/tutorial_basic`.
+
+.. toctree::
+    tutorial_HRB_context
+    tutorial_HRB_implementation
+    tutorial_HRB_results
