+++ log-grid/docs/tutorial/tuto_basic/tutorial_basic.rst
@@ -1,0 +1,8 @@
+###########################
+Tutorial 1: A basic example
+###########################
+
+.. toctree::
+    tutorial_installation
+    tutorial_first_simu
+    tutorial_analyzing_output
