+++ log-grid/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.4]
+
+## Added
+
+- Add HRB tutorial in docs
+
 ## [2.3.3]
 
 ### Fixed

[…]
