### 2024-11-14 #6439 @hippalectryon-0
@slaizet @marlonsmathias @philipcardiff I have added the new example in the docs [here](https://pyloggrid.readthedocs.io/en/joss/tutorial/tuto_HRB/tutorial_HRB.html). Please let me know if this is now satisfactory, and if there's any further point preventing the end of the review.
