+++ paper.bib
@@ -6,7 +6,7 @@
     pages = {154-182},
     year = {1985},
     issn = {0167-2789},
-    doi = {https://doi.org/10.1016/0167-2789(85)90002-8},
+    doi = {10.1016/0167-2789(85)90002-8},
     author = {C. Gloaguen and J. L\'eorat and A. Pouquet and R. Grappin}
 }
 
@@ -78,7 +78,8 @@
   number={6},
   pages={065106},
   year={2023},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/PhysRevE.107.065106}
 }
 
 
@@ -104,6 +105,7 @@ DOI = {10.3390/atmos14111690}
     volume = {280},
     pages = {579--601},
     year = {2015},
-    publisher = {Elsevier}
+    publisher = {Elsevier},
+    doi = {10.1016/j.jcp.2014.09.038}
 }
 
