### 2024-12-09 #6439 @hippalectryon-0
@philipcardiff The references are fixed. The two SKIP doi correspond to Thesis, which do not have doi.
