### 2024-12-09 a4b1ac5
Update paper.md: proposed English edits (#7)
Minor edits to the English.
