### 2024-12-09 77efe0d
update affiliations
