### 2024-12-10 #6439 @hippalectryon-0
@philipcardiff regarding the corresponding author, shouldn't we also add his/her email ? I tried adding an "email" field but it didn't render.
