### 2024-12-10 #6439 @philipcardiff
> * [x]  Double check authors and affiliations (including ORCIDs)
>
> @philipcardiff regarding the corresponding author, shouldn't we also add his/her email ? I tried adding an "email" field but it didn't render.

Hi @hippalectryon-0, you will see from published articles at https://joss.theoj.org that email addresses do not appear on the paper. Instead, the authors' GitHub profile and ORCID links are given.

@kyleniemeyer, feel free to clarify if needed.
