### 2024-12-10 #6439 @hippalectryon-0
The release is [joss-final-1](https://github.com/hippalectryon-0/pyloggrid/releases/tag/joss-final-1)
The doi is [10.5281/zenodo.14356262](https://doi.org/10.5281/zenodo.14356261)

@philipcardiff I think it's all good on my end, can you confirm ? Thanks !
