### 2024-12-10 #6439 @philipcardiff
Thanks, @hippalectryon-0. Normally, the JOSS-specific version would follow the same naming/numbering convention that is adopted in the repository. Would you consider renaming the release as 2.3.2, for example?
