### 2024-12-10 #6439 @hippalectryon-0
> Thanks, @hippalectryon-0. Normally, the JOSS-specific version would follow the same naming/numbering convention that is adopted in the repository. Would you consider renaming the release as 2.3.2, for example?

Sure ! https://github.com/hippalectryon-0/pyloggrid/releases/tag/2.3.2
