### 2024-12-10 #6439 @philipcardiff
Thanks, @hippalectryon-0. Can you update this line in the paper, `This corresponds to version 2.2.1.` to 2.3.2? I assume the text is still relevant for 2.3.2.
