### 2024-12-11 #6439 @kyleniemeyer
@hippalectryon-0 it looks like the version on the Zenodo archive (v1) does not match the version accepted (v2.3.2). Can you correct the archive metadata?

In addition, there are a number of proper names in the references that are not being capitalized properly, likely due to them missing curly brackets to protect in the `title` field (e.g., `{R}unge--{K}utta`). Can you correct all these?
