+++ paper.bib
@@ -1,5 +1,5 @@
 @article{gloaguen85,
-    title = {A scalar model for MHD turbulence},
+    title = {A scalar model for {MHD} turbulence},
     journal = {Physica D: Nonlinear Phenomena},
     volume = {17},
     number = {2},
@@ -60,7 +60,7 @@
 }
 
 @article{barral2023asymptotic,
-    title = {Asymptotic ultimate regime of homogeneous Rayleigh--B{\'e}nard convection on logarithmic lattices},
+    title = {Asymptotic ultimate regime of homogeneous {R}ayleigh--{B}{\'e}nard convection on logarithmic lattices},
     author = {Barral, Amaury and Dubrulle, Berengere},
     journal = {Journal of Fluid Mechanics},
     volume = {962},
@@ -71,7 +71,7 @@
 }
 
 @article{costa2023reversible,
-  title={Reversible Navier-Stokes equation on logarithmic lattices},
+  title={Reversible {N}avier-{S}tokes equation on logarithmic lattices},
   author={Costa, Guillaume and Barral, Amaury and Dubrulle, B{\'e}reng{\`e}re},
   journal={Physical Review E},
   volume={107},
@@ -99,7 +99,7 @@ DOI = {10.3390/atmos14111690}
 
 
 @article{whalen2015exponential,
-    title = {Exponential time-differencing with embedded Runge--Kutta adaptive step control},
+    title = {Exponential time-differencing with embedded {R}unge--{K}utta adaptive step control},
     author = {Whalen, Patrick and Brio, Moysey and Moloney, Jerome V},
     journal = {Journal of Computational Physics},
     volume = {280},
