### 2024-12-11 #6439 @hippalectryon-0
However, the release 2.3.2 now points towards the "old" version - as does the zenodo archive (without the proper references). How should we resolve this ?
