### 2024-12-16 #6439 @kyleniemeyer
Changes to the JOSS paper itself do not need to be reflected in the software version release / archive.
